Add helper functions to QRResult (#23)
* Add helper functions to QRResult

Adding functions to produce Q and R matrices as well as reconstruct the original matrix.

* Fix tabulation

* Removing extra copy

* Adjusting work variable in Q

* Additional tab fixes

* Spacing fixes

* Replace R function with eachLower

* Fix typo

* Making R/reconstruct universal

* Adjusting eachLower in R

The checks not passing is because the cdouble in one of the unittests does not interact well with how eachLower is used. Basically, the 0 in eachLower!"a = 0" does not take into account the cdouble type. I have modified it to include a cast statement.

* Add UT for rectangular qrDecomp

* Updating mir.lapack dependency

* Fix transposed on UT

Likely still going to have some lingering failure...

* Update to fix UTs